requirements.txt
aniso8601==1.1.0
click==6.6
Flask==0.11.1
Flask-RESTful==0.3.5
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pip==8.1.2
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.12
setuptools==26.1.1
six==1.10.0
SQLAlchemy==1.0.15
Werkzeug==0.11.11
wheel==0.29.0
WTForms==2.1
